fix: move isAlternating function to utils
