+++ src/utils/DateTime.ts
@@ -1,4 +1,5 @@
 import { Show } from '@customTypes/RecentlyPlayed';
+import { ScheduleShow } from '@customTypes/Schedule';
 import { debugError } from '@utils/Debug';
 
 export const dayNames = [
@@ -139,4 +140,35 @@ export const formatArchiveDate = (dateString: string) => {
     day: 'numeric',
     year: 'numeric',
   });
+};
+
+export const isAlternatingShowActive = (
+  show: Show | ScheduleShow,
+  targetDate: Date,
+): boolean => {
+  if (show.alternates === 0) {
+    return true; // Non-alternating show is always active
+  }
+
+  // For alternating shows, we need a reference date to calculate weeks.
+  // This should be updated to the first Monday of every season.
+  const referenceDate = new Date('2026-05-25T00:00:00-04:00'); // Eastern Time
+
+  const targetDateEastern = new Date(
+    targetDate.toLocaleString('en-US', { timeZone: 'America/New_York' }),
+  );
+
+  // Calculate weeks since reference date.
+  const daysDiff = Math.floor(
+    (targetDateEastern.getTime() - referenceDate.getTime()) /
+      (1000 * 60 * 60 * 24),
+  );
+
+  const weeksSince = Math.floor(daysDiff / 7);
+
+  if (show.alternates < 3) {
+    return weeksSince % 2 === show.alternates - 1; // 1 means active on odd weeks, 2 means active on even weeks
+  }
+
+  return weeksSince % 4 === show.alternates - 5; // 5 means active on week 1, 6 means active on week 2, etc.
 };
